Move peak finder utils closer
