+++ geoapps/peak_finder/driver.py
@@ -22,13 +22,10 @@ from tqdm import tqdm
 from geoapps.base.application import BaseApplication
 from geoapps.utils import geophysical_systems
 from geoapps.utils.formatters import string_name
-from geoapps.utils.utils import (
-    default_groups_from_property_group,
-    find_anomalies,
-    hex_to_rgb,
-)
+from geoapps.utils.utils import hex_to_rgb
 
 from .params import PeakFinderParams
+from .utils import default_groups_from_property_group, find_anomalies
 
 
 class PeakFinderDriver:
